Update UpgradeContrast.md
No text entry this time. I find it confusing to talk about Upgrading Contrast in the title. It is either the Contrast Application or Team. Server isn't it? Also, I could not manage to find this place with the search like other things. Can enhance the search to find more items?